add vertex buffer to the vulkan renderer

+++ kanji/core/math/vector.h
@@ -6,13 +6,13 @@
 
 //vector 2D
 typedef struct vec2 {
-    double x;
-    double y;
+    float x;
+    float y;
 } vec2;
 
 //vector 3D
 typedef struct vec3 {
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 } vec3;
